+++ src/components/commnunityDetail/Comment.tsx
@@ -32,3 +32,4 @@ export default function Comment({
     </div>
   )
 }
+
